refactor: move functionality to util.py
