+++ aws-python-api/Dockerfile
@@ -67,6 +67,7 @@ COPY templates /usr/src/app/templates
 COPY handler.py /usr/src/app/handler.py
 COPY handler_s3.py /usr/src/app/handler_s3.py
 COPY settings.py /usr/src/app/settings.py
+COPY util.py /usr/src/app/util.py
 WORKDIR /usr/src/app
 
 COPY cred.json /usr/src/app/cred.json
